refactor models, unnecessary file sharding
